+++ CHANGELOG.rst
@@ -10,9 +10,11 @@ Added
 =====
 - Added a UI button for redeploying an EVC.
 - UNI tag_type are now accepted as string.
+- EVCs now listen to ``switch.interface.(link_up|link_down|created|deleted)`` events for activation/deactivation
 
 Changed
 =======
+- EVCs will try to maintain their current_path on link status changes
 - UNIs now will use and free tags from ``Interface.available_tags``.
 - UNI tag_type is changed to string from 1, 2 and 3 values to ``"vlan"``, ``"vlan_qinq"`` and ``"mpls"`` respectively.
 - Add ``set_vlan`` only if UNI A vlan and UNI z vlan are different.